Extract AddrRepr to model/addr.rs
